primer commit

+++ database/migrations/2019_01_21_123424_create_places_table.php
@@ -19,8 +19,8 @@ class CreatePlacesTable extends Migration
             $table->string('description');
             $table->string('startDate');
             $table->string('endDate');
-            $table->float('coordX');
-            $table->float('coordY');
+            $table->double('coordX');
+            $table->double('coordY');
             $table->integer('user_id');
             $table->timestamps();
         });
